requirements.txt
pymongo
fastapi[standard]
python-dotenv
